Make integration tests to be snapshot tests
The tests make sure that the machines in the client/hwlib/test_data
directory are indicated as certified and the hardware details match.
They currently load data for this machines from C3, but we mock C3
responses with the test data from the server/scripts/c3_test_data
directory before the tests run.