License additions (#245)
* The hardware-api server should be AGPLv3 licensed

This project maintains a monorepo with components distributed
in separate packages. The components should use the appropriate
open source licenses.

Note that the charm for the server is Apache Licensed and a
license file already exists in that directory.

* The hardware-api hwlib should be LGPLv3 licensed

This project maintains a monorepo with components distributed
in separate packages. The components should use the appropriate
open source licenses.

* The hardware-api hwctl should be GPLv3 licensed

This project maintains a monorepo with components distributed
in separate packages. The components should use the appropriate
open source licenses.

* Update the files headers in hwapi server side to use GNU AGPLv3

* Update the files headers in hwlib/src/ and debian/ to use GNU LGPLv3

* Fix file headers to use only *GPLv3 license and remove Copyright notice.

---------

Co-authored-by: Nadzeya H <nadzeya.hutsko@canonical.com>