Update client snapcraft.yaml (#322)
* Update snapcraft.yaml

The client snap was missing a homepage, license, specification, and stage packages.

I only added `kmod` as a stage dependency because neither `lsb-release` nor `dpkg` can be used within confinement, so there is no point in staging them for now

* Add more links to the snap metadata

* Upload snapcraft logs and use multiarch build

Still only building for amd64, but this makes adding new platforms easier in the future.

* rename publish step

* Update client/snapcraft.yaml

Co-authored-by: Nadzeya H <84857215+nadzyah@users.noreply.github.com>

---------

Co-authored-by: Nadzeya H <84857215+nadzyah@users.noreply.github.com>