feat(client): change the debian package to non-native format
Since from the practical and conceptual perspective it makes more sense
to change the package format to non-native, the debian/ folder and all
information about the package being available in debian format should be
gone from the upstream. This change keeps only Rust code in the client
side of the repo and removes all the debian-package related information.
Packaging process will be handled using the source package repo:
https://code.launchpad.net/ubuntu/+source/rust-hwlib